fix(suite): fix sending flow mismatch, refactor states
