+++ packages/suite/src/components/suite/modals/ReduxModal/TransactionReviewModal/TransactionReviewModalBody.tsx
@@ -1,0 +1,138 @@
+import { useCallback, useEffect, useState } from 'react';
+
+import { SendState, StakeState, selectSelectedDevice } from '@suite-common/wallet-core';
+import { FormState } from '@suite-common/wallet-types';
+import {
+    constructTransactionReviewOutputsOptional,
+    getTxValidityTimeoutInMs,
+    isRbfBumpFeeTransaction,
+} from '@suite-common/wallet-utils';
+import TrezorConnect from '@trezor/connect';
+import { EventType, analytics } from '@trezor/suite-analytics';
+import { Deferred } from '@trezor/utils';
+
+import { useSelector } from 'src/hooks/suite';
+import { selectAccountIncludingChosenInTrading } from 'src/reducers/wallet/selectedAccountReducer';
+import { redactRouterUrl } from 'src/utils/suite/analytics';
+
+import { isStakeState } from './TransactionReviewModal';
+import { TransactionReviewModalBodyInner } from './TransactionReviewModalBodyInner';
+import { ConfirmActionModal } from '../DeviceContextModal/ConfirmActionModal';
+
+export const hasTxValidityExpired = (deadline: number) => deadline <= Date.now();
+
+export const getTxType = (txInfoState: SendState | StakeState, precomposedForm: FormState) => {
+    const stakeType = isStakeState(txInfoState) ? 'stake' : undefined;
+    const tradeType = precomposedForm.trading?.activeSection ? 'trade' : undefined;
+
+    return stakeType ?? tradeType;
+};
+
+export type TransactionReviewModalBodyProps = {
+    decision: Deferred<boolean, string | number | undefined> | undefined;
+    txInfoState: SendState | StakeState;
+    tryAgainSignTx: () => void;
+    cancelSignTx: () => void;
+    isRbfConfirmedError?: boolean;
+    precomposedForm?: FormState;
+};
+
+export const TransactionReviewModalBody = ({
+    decision,
+    txInfoState,
+    tryAgainSignTx,
+    cancelSignTx,
+    precomposedForm,
+    isRbfConfirmedError,
+}: TransactionReviewModalBodyProps) => {
+    const account = useSelector(selectAccountIncludingChosenInTrading);
+    const device = useSelector(selectSelectedDevice);
+    const [isSending, setIsSending] = useState(false);
+    const { precomposedTx } = txInfoState;
+    const [hasTxExpired, setHasTxExpired] = useState(false);
+
+    const router = useSelector(state => state.router);
+
+    const createdTxTimestamp = txInfoState?.precomposedTx?.createdTimestamp ?? 0;
+    const shouldCheckTxTimeValidity = account?.networkType === 'solana' && createdTxTimestamp !== 0;
+
+    const deadline = createdTxTimestamp + getTxValidityTimeoutInMs(account?.networkType);
+
+    useEffect(() => {
+        if (!shouldCheckTxTimeValidity) {
+            return;
+        }
+
+        const now = Date.now();
+        const timeLeft = Math.max(deadline - now, 0);
+        let mounted = true;
+
+        const timeoutId = setTimeout(() => {
+            if (mounted && !isSending) {
+                setHasTxExpired(true);
+                TrezorConnect.cancel('tx-timeout');
+            }
+        }, timeLeft);
+
+        return () => {
+            mounted = false;
+            clearTimeout(timeoutId);
+        };
+    }, [deadline, isSending, shouldCheckTxTimeValidity]);
+
+    const isBumpFeeRbfAction =
+        precomposedTx !== undefined && isRbfBumpFeeTransaction(precomposedTx);
+
+    const decreaseOutputId =
+        isBumpFeeRbfAction && precomposedTx.useNativeRbf
+            ? precomposedForm?.setMaxOutputId
+            : undefined;
+
+    const outputs = constructTransactionReviewOutputsOptional({
+        account,
+        decreaseOutputId,
+        device,
+        precomposedForm,
+        precomposedTx,
+    });
+
+    const handleTryAgain = useCallback(
+        (cancel: boolean) => {
+            if (cancel) {
+                TrezorConnect.cancel('tx-timeout');
+            }
+
+            tryAgainSignTx();
+
+            analytics.report({
+                type: EventType.TransactionRetry,
+                payload: { url: redactRouterUrl(router.url) },
+            });
+        },
+        [tryAgainSignTx, router.url],
+    );
+
+    if (!device) return null;
+    if (!account || !precomposedTx || !precomposedForm) {
+        // TODO: special case for Connect Popup
+        return <ConfirmActionModal device={device} />;
+    }
+
+    return (
+        <TransactionReviewModalBodyInner
+            account={account}
+            outputs={outputs}
+            decision={decision}
+            txInfoState={txInfoState}
+            tryAgainSignTx={tryAgainSignTx}
+            cancelSignTx={cancelSignTx}
+            precomposedForm={precomposedForm}
+            precomposedTx={precomposedTx}
+            isSending={isSending}
+            setIsSending={setIsSending}
+            handleTryAgain={handleTryAgain}
+            hasTxExpired={hasTxExpired}
+            isRbfConfirmedError={isRbfConfirmedError}
+        />
+    );
+};
