--- packages/suite/src/components/suite/modals/ReduxModal/TransactionReviewModal/TransactionReviewModalContent.tsx
@@ -1,474 +1,0 @@
-import { useEffect, useState } from 'react';
-
-import { selectConnectPopupCall } from '@suite-common/connect-popup';
-import { notificationsActions } from '@suite-common/toast-notifications';
-import { selectTradingComposedTransactionInfo } from '@suite-common/trading';
-import { NetworkType } from '@suite-common/wallet-config';
-import {
-    DeviceRootState,
-    SendState,
-    SerializedTx,
-    StakeState,
-    selectIsTxOutputInternal,
-    selectSelectedDevice,
-    selectSendFormReviewButtonRequestsCount,
-} from '@suite-common/wallet-core';
-import { FormState, RbfTransactionType, StakeType } from '@suite-common/wallet-types';
-import {
-    constructTransactionReviewOutputsOptional,
-    getStakeType,
-    isRbfBumpFeeTransaction,
-    isRbfCancelTransaction,
-    isRbfTransaction,
-} from '@suite-common/wallet-utils';
-import { Column, Modal, Row } from '@trezor/components';
-import TrezorConnect from '@trezor/connect';
-import { copyToClipboard, download } from '@trezor/dom-utils';
-import { ConfirmOnDevice } from '@trezor/product-components';
-import { EventType, TransactionCreatedEvent, analytics } from '@trezor/suite-analytics';
-import { spacings } from '@trezor/theme';
-import { Deferred } from '@trezor/utils';
-
-import * as modalActions from 'src/actions/suite/modalActions';
-import { Translation } from 'src/components/suite';
-import { ConnectModalBackdrop } from 'src/components/suite/ConnectModalBackdrop';
-import { useDispatch, useSelector } from 'src/hooks/suite';
-import { selectIsActionAbortable } from 'src/reducers/suite/suiteReducer';
-import { selectAccountIncludingChosenInTrading } from 'src/reducers/wallet/selectedAccountReducer';
-import { redactRouterUrl } from 'src/utils/suite/analytics';
-import { getTransactionReviewModalActionTranslation } from 'src/utils/suite/transactionReview';
-
-import { TransactionReviewDetails } from './TransactionReviewDetails';
-import { isStakeState } from './TransactionReviewModal';
-import { TransactionReviewOutputList } from './TransactionReviewOutputList/TransactionReviewOutputList';
-import { TransactionReviewOutputTimer } from './TransactionReviewOutputList/TransactionReviewOutputTimer';
-import { TransactionReviewSummary } from './TransactionReviewSummary';
-import { ConfirmActionModal } from '../DeviceContextModal/ConfirmActionModal';
-import { ExpiredTxValidity } from '../UserContextModal/TxDetailModal/ExpiredTxValidity';
-import { ReplaceByFeeFailedOriginalTxConfirmed } from '../UserContextModal/TxDetailModal/ReplaceByFeeFailedOriginalTxConfirmed';
-
-const getTxValidityTimeoutInMs = (networkType?: NetworkType) => {
-    if (networkType === 'solana') {
-        // Blockhash required in Solana tx is valid for 1 minute. Leave 15 seconds for tx confirmation.
-        return 45 * 1000;
-    }
-
-    return 0;
-};
-
-const hasTxValidityExpired = (deadline: number) => deadline <= Date.now();
-
-type ShouldShowTxValidityTimerProps = {
-    deadline: number;
-    buttonRequestsCount: number;
-    serializedTx: SerializedTx | undefined;
-    stakeType: StakeType | null;
-    shouldCheckTxTimeValidity: boolean;
-    isInternalTransfer: boolean;
-    isTrading: boolean;
-};
-
-const shouldShowTxValidityTimer = ({
-    deadline,
-    buttonRequestsCount,
-    serializedTx,
-    stakeType,
-    shouldCheckTxTimeValidity,
-    isInternalTransfer,
-    isTrading,
-}: ShouldShowTxValidityTimerProps) => {
-    if (!shouldCheckTxTimeValidity || hasTxValidityExpired(deadline)) {
-        return false;
-    }
-
-    const isFirstStep = buttonRequestsCount <= 1;
-    const isStaking = stakeType && !serializedTx;
-
-    return isInternalTransfer || !isFirstStep || serializedTx || isStaking || isTrading;
-};
-
-const getTxType = (txInfoState: SendState | StakeState, precomposedForm: FormState) => {
-    const stakeType = isStakeState(txInfoState) ? 'stake' : undefined;
-    const tradeType = precomposedForm.trading?.activeSection ? 'trade' : undefined;
-
-    return stakeType ?? tradeType;
-};
-
-const mapRbfTypeToReporting: Record<
-    RbfTransactionType,
-    TransactionCreatedEvent['payload']['action']
-> = { 'bump-fee': 'replaced', cancel: 'canceled' };
-
-type TransactionReviewModalContentProps = {
-    decision: Deferred<boolean, string | number | undefined> | undefined;
-    txInfoState: SendState | StakeState;
-    tryAgainSignTx: () => void;
-    cancelSignTx: () => void;
-    isRbfConfirmedError?: boolean;
-    precomposedForm?: FormState;
-};
-
-export const TransactionReviewModalContent = ({
-    decision,
-    txInfoState,
-    tryAgainSignTx,
-    cancelSignTx,
-    isRbfConfirmedError,
-    precomposedForm,
-}: TransactionReviewModalContentProps) => {
-    const dispatch = useDispatch();
-    const account = useSelector(selectAccountIncludingChosenInTrading);
-    const device = useSelector(selectSelectedDevice);
-    const isActionAbortable = useSelector(selectIsActionAbortable);
-    const connectPopupCall = useSelector(selectConnectPopupCall);
-    const [isSending, setIsSending] = useState(false);
-    const [areDetailsVisible, setAreDetailsVisible] = useState(false);
-    const deviceModelInternal = device?.features?.internal_model;
-    const { precomposedTx, serializedTx } = txInfoState;
-    const [hasTxExpired, setHasTxExpired] = useState(false);
-
-    const router = useSelector(state => state.router);
-
-    const tradingToken = useSelector(selectTradingComposedTransactionInfo).composed?.token;
-
-    const createdTxTimestamp = txInfoState?.precomposedTx?.createdTimestamp ?? 0;
-    const shouldCheckTxTimeValidity = account?.networkType === 'solana' && createdTxTimestamp !== 0;
-
-    const deadline = createdTxTimestamp + getTxValidityTimeoutInMs(account?.networkType);
-
-    // check if transaction is still valid
-    useEffect(() => {
-        if (!shouldCheckTxTimeValidity) {
-            return;
-        }
-
-        const now = Date.now();
-        const timeLeft = Math.max(deadline - now, 0);
-        let mounted = true;
-
-        const timeoutId = setTimeout(() => {
-            if (mounted && !isSending) {
-                setHasTxExpired(true);
-                TrezorConnect.cancel('tx-timeout');
-            }
-        }, timeLeft);
-
-        return () => {
-            mounted = false;
-            clearTimeout(timeoutId);
-        };
-    }, [deadline, isSending, shouldCheckTxTimeValidity]);
-
-    const isBumpFeeRbfAction =
-        precomposedTx !== undefined && isRbfBumpFeeTransaction(precomposedTx);
-
-    const decreaseOutputId =
-        isBumpFeeRbfAction && precomposedTx.useNativeRbf
-            ? precomposedForm?.setMaxOutputId
-            : undefined;
-
-    const buttonRequestsCount = useSelector((state: DeviceRootState) =>
-        selectSendFormReviewButtonRequestsCount(state, account?.symbol, decreaseOutputId),
-    );
-
-    const outputs = constructTransactionReviewOutputsOptional({
-        account,
-        decreaseOutputId,
-        device,
-        precomposedForm,
-        precomposedTx,
-    });
-
-    const isInternalTransfer = useSelector(state =>
-        selectIsTxOutputInternal(state, account?.symbol, outputs[0]),
-    );
-
-    if (!device) return null;
-    if (!account || !precomposedTx || !precomposedForm) {
-        // TODO: special case for Connect Popup
-        return <ConfirmActionModal device={device} />;
-    }
-
-    const { symbol, networkType } = account;
-    const { options, selectedFee } = precomposedForm;
-
-    const txType = getTxType(txInfoState, precomposedForm);
-
-    // for bump fee we have to analyze tx data which are in outputs[0], for legacy in outputs[1]
-    const stakeType = getStakeType(precomposedForm, outputs);
-
-    const onCancel = () => {
-        dispatch(modalActions.onCancel());
-
-        if (isActionAbortable || serializedTx) {
-            cancelSignTx();
-            decision?.resolve(false);
-        }
-    };
-
-    const isCancelRbfAction = isRbfCancelTransaction(precomposedTx);
-
-    const isTxExpired = hasTxValidityExpired(deadline);
-
-    const showTxValidityTimer = shouldShowTxValidityTimer({
-        deadline,
-        buttonRequestsCount,
-        serializedTx,
-        stakeType,
-        shouldCheckTxTimeValidity,
-        isInternalTransfer,
-        isTrading: !!precomposedForm.trading?.activeSection,
-    });
-
-    const actionTranslation = getTransactionReviewModalActionTranslation({
-        stakeType,
-        precomposedForm,
-        tradingToken,
-        isBumpFeeRbfAction,
-        isCancelRbfAction,
-        isSending,
-    });
-
-    const isBroadcastEnabled = options.includes('broadcast');
-
-    const reportTransactionCreatedEvent = (action: TransactionCreatedEvent['payload']['action']) =>
-        analytics.report({
-            type: EventType.TransactionCreated,
-            payload: {
-                action,
-                symbol,
-                tokens: outputs
-                    .filter(output => output.token?.symbol)
-                    .map(output => output.token?.symbol)
-                    .join(','),
-                outputsCount: precomposedForm.outputs.length,
-                broadcast: isBroadcastEnabled,
-                bitcoinLocktime: !!options.includes('bitcoinLocktime'),
-                ethereumData: !!options.includes('ethereumData'),
-                ethereumNonce: !!options.includes('ethereumNonce'),
-                destinationTag: !!options.includes('destinationTag'),
-                selectedFee: selectedFee || 'normal',
-                isCoinControlEnabled: precomposedForm.isCoinControlEnabled,
-                hasCoinControlBeenOpened: precomposedForm.hasCoinControlBeenOpened,
-                txType,
-            },
-        });
-
-    const handleSend = () => {
-        if (networkType === 'solana' || networkType === 'stellar') {
-            setIsSending(true);
-        }
-
-        if (decision) {
-            decision.resolve(true);
-            reportTransactionCreatedEvent(
-                isRbfTransaction(precomposedTx)
-                    ? mapRbfTypeToReporting[precomposedTx.rbfType]
-                    : 'sent',
-            );
-
-            if (stakeType) {
-                return analytics.report({
-                    type: EventType.StakingConfirm,
-                    payload: { action: stakeType, networkSymbol: symbol },
-                });
-            }
-        }
-    };
-
-    const handleCopy = () => {
-        const result = copyToClipboard(serializedTx!.tx);
-
-        if (typeof result !== 'string') {
-            dispatch(notificationsActions.addToast({ type: 'copy-to-clipboard' }));
-        }
-
-        reportTransactionCreatedEvent('copied');
-    };
-
-    const handleDownload = () => {
-        download(serializedTx!.tx, 'signed-transaction.txt');
-        reportTransactionCreatedEvent('downloaded');
-    };
-
-    const handleTryAgain = (cancel: boolean) => {
-        if (cancel) {
-            TrezorConnect.cancel('tx-timeout');
-        }
-
-        tryAgainSignTx();
-
-        analytics.report({
-            type: EventType.TransactionRetry,
-            payload: { url: redactRouterUrl(router.url) },
-        });
-    };
-
-    const handleDetailsClick = () => {
-        setAreDetailsVisible(areVisible => {
-            if (!areVisible) {
-                analytics.report({
-                    type: EventType.SendDetailOpened,
-                    payload: {
-                        assetSymbol: symbol,
-                    },
-                });
-            }
-
-            return !areVisible;
-        });
-    };
-
-    const BottomContent = () => {
-        if (isRbfConfirmedError) {
-            return (
-                <Modal.Button variant="tertiary" onClick={onCancel}>
-                    <Translation id="TR_CLOSE" />
-                </Modal.Button>
-            );
-        }
-
-        if (shouldCheckTxTimeValidity && isTxExpired && !isSending) {
-            return (
-                <>
-                    <Modal.Button variant="primary" onClick={() => handleTryAgain(false)}>
-                        <Translation id="TR_TRY_AGAIN" />
-                    </Modal.Button>
-                    <Modal.Button variant="tertiary" onClick={onCancel}>
-                        <Translation id="TR_CLOSE" />
-                    </Modal.Button>
-                </>
-            );
-        }
-
-        if (areDetailsVisible) {
-            return null;
-        }
-
-        if (connectPopupCall?.state === 'ongoing') {
-            return null;
-        }
-
-        if (isBroadcastEnabled) {
-            return (
-                <Modal.Button
-                    data-testid="@modal/send"
-                    isDisabled={!serializedTx || hasTxExpired}
-                    isLoading={isSending}
-                    variant={isCancelRbfAction ? 'destructive' : 'primary'}
-                    onClick={handleSend}
-                >
-                    <Translation {...actionTranslation} />
-                </Modal.Button>
-            );
-        }
-
-        return (
-            <>
-                <Modal.Button
-                    isDisabled={!serializedTx}
-                    onClick={handleCopy}
-                    data-testid="@send/copy-raw-transaction"
-                >
-                    <Translation id="COPY_TRANSACTION_TO_CLIPBOARD" />
-                </Modal.Button>
-                <Modal.Button
-                    variant="tertiary"
-                    isDisabled={!serializedTx}
-                    onClick={handleDownload}
-                >
-                    <Translation id="DOWNLOAD_TRANSACTION" />
-                </Modal.Button>
-            </>
-        );
-    };
-
-    const Content = () => {
-        if (areDetailsVisible) {
-            return <TransactionReviewDetails tx={precomposedTx} txHash={serializedTx?.tx} />;
-        }
-
-        if (isRbfConfirmedError && isRbfTransaction(precomposedTx)) {
-            return (
-                <ReplaceByFeeFailedOriginalTxConfirmed
-                    type={precomposedTx.rbfType}
-                    networkType={networkType}
-                />
-            );
-        }
-
-        if (shouldCheckTxTimeValidity && isTxExpired && !isSending) {
-            return <ExpiredTxValidity symbol={symbol} />;
-        }
-
-        return (
-            <Column gap={spacings.md}>
-                <TransactionReviewOutputList
-                    account={account}
-                    precomposedTx={precomposedTx}
-                    precomposedForm={precomposedForm}
-                    signedTx={serializedTx}
-                    outputs={outputs}
-                    buttonRequestsCount={buttonRequestsCount}
-                    isRbfAction={isBumpFeeRbfAction}
-                    tradingFormState={precomposedForm?.trading}
-                    isSending={isSending}
-                    stakeType={stakeType || undefined}
-                    deadline={deadline}
-                    onTryAgain={handleTryAgain}
-                />
-            </Column>
-        );
-    };
-
-    return (
-        <ConnectModalBackdrop canSwitchDevice>
-            {!isRbfConfirmedError && (
-                <ConfirmOnDevice
-                    title={<Translation id="TR_CONFIRM_ON_TREZOR" />}
-                    steps={outputs.length + 1}
-                    activeStep={serializedTx ? outputs.length + 2 : buttonRequestsCount}
-                    deviceModelInternal={deviceModelInternal}
-                    deviceUnitColor={device?.features?.unit_color}
-                    successText={<Translation id="TR_CONFIRMED_TX" />}
-                    onCancel={isSending ? undefined : onCancel}
-                />
-            )}
-            <Modal.ModalBase
-                heading={
-                    <Translation
-                        {...(areDetailsVisible ? { id: 'TR_DETAIL' } : actionTranslation)}
-                    />
-                }
-                onBackClick={areDetailsVisible ? () => setAreDetailsVisible(false) : undefined}
-                description={
-                    !areDetailsVisible && (
-                        <Row justifyContent="space-between">
-                            <TransactionReviewSummary
-                                tx={precomposedTx}
-                                account={account}
-                                broadcast={isBroadcastEnabled}
-                                onDetailsClick={handleDetailsClick}
-                                stakeType={stakeType}
-                            />
-                            {showTxValidityTimer && (
-                                <Row gap={spacings.xs}>
-                                    <TransactionReviewOutputTimer
-                                        deadline={deadline}
-                                        onTryAgain={handleTryAgain}
-                                        isMinimal
-                                        isSending={isSending}
-                                    />
-                                </Row>
-                            )}
-                        </Row>
-                    )
-                }
-                bottomContent={<BottomContent />}
-                size="small"
-            >
-                <Content />
-            </Modal.ModalBase>
-        </ConnectModalBackdrop>
-    );
-};
