+++ packages/suite/src/components/suite/modals/ReduxModal/TransactionReviewModal/TransactionReviewOutputList/TransactionReviewModalContent.tsx
@@ -1,0 +1,128 @@
+import { useMemo } from 'react';
+import { useSelector } from 'react-redux';
+
+import {
+    DeviceRootState,
+    SerializedTx,
+    selectSelectedDevice,
+    selectSendFormReviewButtonRequestsCount,
+} from '@suite-common/wallet-core';
+import { Account, FormState, GeneralPrecomposedTransactionFinal } from '@suite-common/wallet-types';
+import {
+    constructTransactionReviewOutputsOptional,
+    getStakeType,
+    getTxValidityTimeoutInMs,
+    isRbfBumpFeeTransaction,
+    isRbfTransaction,
+} from '@suite-common/wallet-utils';
+import { Column } from '@trezor/components';
+import { spacings } from '@trezor/theme';
+
+import { ExpiredTxValidity } from '../../UserContextModal/TxDetailModal/ExpiredTxValidity';
+import { ReplaceByFeeFailedOriginalTxConfirmed } from '../../UserContextModal/TxDetailModal/ReplaceByFeeFailedOriginalTxConfirmed';
+import { TransactionReviewDetails } from '../TransactionReviewDetails';
+import { TransactionReviewOutputList } from './TransactionReviewOutputList';
+import { hasTxValidityExpired } from '../TransactionReviewModalBody';
+
+type TransactionReviewModalContentProps = {
+    account: Account;
+    precomposedTx: GeneralPrecomposedTransactionFinal;
+    precomposedForm: FormState;
+    isSending: boolean;
+    onTryAgain: (cancel: boolean) => void;
+    reviewStep: number;
+    serializedTx?: SerializedTx;
+    areDetailsVisible?: boolean;
+    isRbfConfirmedError?: boolean;
+};
+
+export const TransactionReviewModalContent = ({
+    account,
+    precomposedTx,
+    areDetailsVisible,
+    serializedTx,
+    reviewStep,
+    precomposedForm,
+    onTryAgain,
+    isSending,
+    isRbfConfirmedError,
+}: TransactionReviewModalContentProps) => {
+    const { symbol, networkType } = account;
+    const device = useSelector(selectSelectedDevice);
+
+    const createdTxTimestamp = useMemo(
+        () => precomposedTx.createdTimestamp ?? 0,
+        [precomposedTx.createdTimestamp],
+    );
+
+    const deadline = createdTxTimestamp + getTxValidityTimeoutInMs(account?.networkType);
+    const isTxExpired = hasTxValidityExpired(deadline);
+
+    const isBumpFeeRbfAction =
+        precomposedTx !== undefined && isRbfBumpFeeTransaction(precomposedTx);
+
+    const decreaseOutputId =
+        isBumpFeeRbfAction && precomposedTx.useNativeRbf
+            ? precomposedForm?.setMaxOutputId
+            : undefined;
+
+    const buttonRequestsCount = useSelector((state: DeviceRootState) =>
+        selectSendFormReviewButtonRequestsCount(state, symbol, decreaseOutputId),
+    );
+
+    const outputs = useMemo(
+        () =>
+            constructTransactionReviewOutputsOptional({
+                account,
+                decreaseOutputId,
+                device,
+                precomposedForm,
+                precomposedTx,
+            }),
+        [account, decreaseOutputId, device, precomposedForm, precomposedTx],
+    );
+
+    const stakeType = getStakeType(precomposedForm, outputs);
+
+    const shouldCheckTxTimeValidity = useMemo(
+        () => account.networkType === 'solana' && createdTxTimestamp !== 0,
+        [account.networkType, createdTxTimestamp],
+    );
+
+    if (areDetailsVisible) {
+        return <TransactionReviewDetails tx={precomposedTx} txHash={serializedTx?.tx} />;
+    }
+
+    if (isRbfConfirmedError && isRbfTransaction(precomposedTx)) {
+        return (
+            <ReplaceByFeeFailedOriginalTxConfirmed
+                type={precomposedTx.rbfType}
+                networkType={networkType}
+            />
+        );
+    }
+
+    if (shouldCheckTxTimeValidity && isTxExpired && !isSending) {
+        return <ExpiredTxValidity symbol={symbol} />;
+    }
+
+    return (
+        <Column gap={spacings.md}>
+            <TransactionReviewOutputList
+                account={account}
+                precomposedTx={precomposedTx}
+                precomposedForm={precomposedForm}
+                signedTx={serializedTx}
+                outputs={outputs}
+                buttonRequestsCount={buttonRequestsCount}
+                isRbfAction={isBumpFeeRbfAction}
+                tradingFormState={precomposedForm?.trading}
+                reviewStep={reviewStep}
+                isSending={isSending}
+                stakeType={stakeType || undefined}
+                deadline={deadline}
+                onTryAgain={onTryAgain}
+            />
+        </Column>
+    );
+};
