+++ packages/suite/src/components/suite/modals/ReduxModal/TransactionReviewModal/TransactionReviewOutputList/__tests__/getTransactionReviewState.test.ts
@@ -1,34 +1,63 @@
 import { getTransactionReviewState } from '../getTransactionReviewState';
 
-describe('Transaction review states', () => {
-    it('should return "confirmed" state for signed transactions', () => {
-        const state = getTransactionReviewState(1, 0, true, 'ButtonRequest_SignTx');
+describe(getTransactionReviewState.name, () => {
+    it('returns "confirmed" state for signed transactions', () => {
+        const state = getTransactionReviewState({
+            index: 1,
+            currentStep: 0,
+            hasSignedTx: true,
+            lastButtonRequestCode: 'ButtonRequest_SignTx',
+        });
         expect(state).toBe('confirmed');
     });
 
-    it('should return "unconfirmed" state for unsigned transactions', () => {
-        const state = getTransactionReviewState(1, 0, false, null);
+    it('returns "unconfirmed" state for unsigned transactions', () => {
+        const state = getTransactionReviewState({
+            index: 1,
+            currentStep: 0,
+            hasSignedTx: false,
+        });
         expect(state).toBe('unconfirmed');
     });
 
-    it('should return "active" state for the current step', () => {
-        const state = getTransactionReviewState(1, 1, false, null);
+    it('returns "active" state for the current step', () => {
+        const state = getTransactionReviewState({
+            index: 1,
+            currentStep: 1,
+            hasSignedTx: false,
+        });
         expect(state).toBe('active');
     });
 
-    it('should return "unconfirmed" state for steps beyond the current step', () => {
-        const state = getTransactionReviewState(2, 3, false, null);
+    it('returns "unconfirmed" state for steps beyond the current step', () => {
+        const state = getTransactionReviewState({
+            index: 4,
+            currentStep: 2,
+            hasSignedTx: false,
+        });
         expect(state).toBe('unconfirmed');
     });
 
-    it('should correctly determine output states based on review step', () => {
-        const outputState = getTransactionReviewState(0, 1, false);
+    it('correctly determines output states based on review step', () => {
+        const outputState = getTransactionReviewState({
+            index: 0,
+            currentStep: 1,
+            hasSignedTx: false,
+        });
         expect(outputState).toBe('confirmed');
 
-        const outputState2 = getTransactionReviewState(1, 1, false);
+        const outputState2 = getTransactionReviewState({
+            index: 1,
+            currentStep: 1,
+            hasSignedTx: false,
+        });
         expect(outputState2).toBe('active');
 
-        const outputState3 = getTransactionReviewState(2, 1, false);
+        const outputState3 = getTransactionReviewState({
+            index: 2,
+            currentStep: 1,
+            hasSignedTx: false,
+        });
         expect(outputState3).toBe('unconfirmed');
     });
 });
